backend/requirements.txt
Django==5.0.2
djangorestframework==3.15.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
duckdb==0.9.2
gunicorn==21.2.0
python-magic==0.4.27
langchain==0.1.14
langgraph==0.0.14 